+++ src/i18n/langs/zh.json
@@ -724,6 +724,8 @@
     "CreateReplayStorage": "创建录像存储",
     "Dashboard": "仪表盘",
     "DatabaseApp": "数据库",
+    "RemoteApps": "远程应用",
+    "Clouds": "云设施",
     "DatabaseAppCreate": "创建数据库应用",
     "DatabaseAppDetail": "数据库详情",
     "DatabaseAppPermission": "数据库授权",
