+++ src/router/console/assets.js
@@ -46,111 +46,6 @@ export default [
     ]
   },
   {
-    path: 'hosts',
-    component: empty,
-    redirect: '',
-    meta: { title: i18n.t('route.AssetList'), permissions: ['assets.view_asset'] },
-    children: [
-      {
-        path: '',
-        name: 'HostList',
-        component: () => import('@/views/assets/Host/HostList.vue'),
-        meta: { title: i18n.t('route.HostList') }
-      },
-      {
-        path: 'create',
-        name: 'HostCreate',
-        component: () => import('@/views/assets/Asset/AssetCreateUpdate.vue'),
-        hidden: true,
-        meta: { title: i18n.t('route.HostCreate') }
-      },
-      {
-        path: ':id',
-        name: 'HostDetail',
-        component: () => import('@/views/assets/Asset/AssetDetail'),
-        hidden: true,
-        meta: { title: i18n.t('route.HostDetail') }
-      },
-      {
-        path: ':id/update',
-        name: 'HostUpdate',
-        component: () => import('@/views/assets/Asset/AssetCreateUpdate.vue'),
-        hidden: true,
-        meta: { title: i18n.t('route.HostUpdate') }
-      }
-    ]
-  },
-  {
-    path: 'devices',
-    component: empty,
-    redirect: '',
-    meta: { title: i18n.t('route.NetworkDevices'), permissions: ['assets.view_asset'] },
-    children: [
-      {
-        path: '',
-        name: 'DeviceList',
-        component: () => import('@/views/assets/Asset/AssetList.vue'),
-        meta: { title: i18n.t('route.NetworkDevices'), permissions: [] }
-      },
-      {
-        path: 'create',
-        name: 'DeviceCreate',
-        component: () => import('@/views/assets/Asset/AssetCreateUpdate.vue'),
-        hidden: true,
-        meta: { title: i18n.t('route.DeviceCreate') }
-      },
-      {
-        path: ':id',
-        name: 'DeviceDetail',
-        component: () => import('@/views/assets/Asset/AssetDetail'),
-        hidden: true,
-        meta: { title: i18n.t('route.DeviceDetail') }
-      },
-      {
-        path: ':id/update',
-        name: 'DeviceUpdate',
-        component: () => import('@/views/assets/Asset/AssetCreateUpdate.vue'),
-        hidden: true,
-        meta: { title: i18n.t('route.DeviceUpdate') }
-      }
-    ]
-  },
-  {
-    path: 'databases',
-    component: empty,
-    redirect: '',
-    meta: { title: i18n.t('route.Databases'), permissions: [] },
-    children: [
-      {
-        path: '',
-        name: 'DatabaseList',
-        component: () => import('@/views/assets/Asset/AssetList.vue'),
-        meta: { title: i18n.t('route.Databases'), permissions: [] }
-      },
-      {
-        path: 'create',
-        name: 'DatabaseCreate',
-        component: () => import('@/views/assets/Asset/AssetCreateUpdate.vue'),
-        hidden: true,
-        meta: { title: i18n.t('route.DatabaseCreate') }
-      },
-      {
-        path: ':id',
-        name: 'AssetDetail',
-        component: () => import('@/views/assets/Asset/AssetDetail'),
-        hidden: true,
-        meta: { title: i18n.t('route.AssetDetail') }
-      },
-      {
-        path: ':id/update',
-        name: 'AssetUpdate',
-        component: () => import('@/views/assets/Asset/AssetCreateUpdate.vue'),
-        hidden: true,
-        meta: { title: i18n.t('route.AssetUpdate') }
-      }
-    ]
-  },
-  {
     path: 'domains',
     component: empty,
     redirect: '',
